requirements.txt
python-dotenv~=1.0.0
python-telegram-bot
openai

pytest
pytest-cov
pytest-asyncio

# for lambda
boto3
